corrections
git-svn-id: https://sbia-svn.uphs.upenn.edu/projects/CBICA_Toolkit/trunk@336 36980a1d-918e-4cb2-ab0a-7661cd27ea28

+++ src/classes/cbicaUtilities.cpp
@@ -990,7 +990,7 @@ namespace cbica
     int threads = omp_get_max_threads(); // obtain maximum number of threads available on machine  
     // if the total number of rows in CSV file are less than the available number of threads on machine (happens for testing),
     // use only the number of rows where meaningful data is present - this avoids extra thread overhead
-    threads > numberOfRows ? threads = static_cast<int>(numberOfRows - 1) : threads = threads; 
+    threads > static_cast<int>(numberOfRows) ? threads = static_cast<int>(numberOfRows - 1) : threads = threads; 
 #pragma omp parallel for num_threads(threads)
     for (int rowCounter = 1; rowCounter < allRows.size(); rowCounter++)
     {
